add initial_condition.py
