+++ flow360/component/flow360_params/flow360_params.py
@@ -72,6 +72,7 @@ from .flow360_output import (
     VolumeOutput,
     VolumeOutputLegacy,
 )
+from .initial_condition import InitialConditions
 from .params_base import (
     DeprecatedAlias,
     Flow360BaseModel,
@@ -756,32 +757,6 @@ class USstandardAtmosphere(Flow360BaseModel):
 air = AirDensityTemperature(temperature=288.15 * u.K, density=1.225 * u.kg / u.m**3)
 
 FluidPropertyTypes = Union[AirDensityTemperature, AirPressureTemperature]
-
-
-class InitialCondition(Flow360BaseModel):
-    """:class:`InitialCondition` class"""
-
-    type: str
-
-
-class FreestreamInitialCondition(InitialCondition):
-    """:class:`FreestreamInitialCondition` class"""
-
-    type: Literal["freestream"] = pd.Field("freestream", const=True)
-
-
-class ExpressionInitialCondition(InitialCondition):
-    """:class:`ExpressionInitialCondition` class"""
-
-    type: Literal["expression"] = pd.Field("expression", const=True)
-    rho: str = pd.Field()
-    u: str = pd.Field()
-    v: str = pd.Field()
-    w: str = pd.Field()
-    p: str = pd.Field()
-
-
-InitialConditions = Union[FreestreamInitialCondition, ExpressionInitialCondition]
 
 
 class BETDiskTwist(Flow360BaseModel):
