+++ flow360/component/flow360_params/initial_condition.py
@@ -1,0 +1,37 @@
+"""
+Initial condition parameters
+"""
+from __future__ import annotations
+
+from typing import Union
+
+import pydantic as pd
+from typing_extensions import Literal
+
+from .params_base import Flow360BaseModel
+
+
+class InitialCondition(Flow360BaseModel):
+    """:class:`InitialCondition` class"""
+
+    type: str
+
+
+class FreestreamInitialCondition(InitialCondition):
+    """:class:`FreestreamInitialCondition` class"""
+
+    type: Literal["freestream"] = pd.Field("freestream", const=True)
+
+
+class ExpressionInitialCondition(InitialCondition):
+    """:class:`ExpressionInitialCondition` class"""
+
+    type: Literal["expression"] = pd.Field("expression", const=True)
+    rho: str = pd.Field()
+    u: str = pd.Field()
+    v: str = pd.Field()
+    w: str = pd.Field()
+    p: str = pd.Field()
+
+
+InitialConditions = Union[FreestreamInitialCondition, ExpressionInitialCondition]
