+++ tests/params/test_initial_condition.py
@@ -2,7 +2,7 @@ import unittest
 
 import pytest
 
-from flow360.component.flow360_params.flow360_params import (
+from flow360.component.flow360_params.initial_condition import (
     ExpressionInitialCondition,
     FreestreamInitialCondition,
 )
